using mongoose to connect to the database
